requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-language==1.3.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
